book-scan-ocr/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
setuptools>=69.0.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
aiofiles==24.1.0

# PDF & Image processing
pdf2image==1.17.0
opencv-python-headless==4.10.0.84
Pillow==10.4.0

# OCR
easyocr>=1.7.1
openai==1.45.0
anthropic==0.34.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
